added farthest north 4

+++ lib/src/main/java/edu/illinois/cs/cs125/spring2019/mp0/lib/Locator.java
@@ -54,7 +54,7 @@ public class Locator {
                 if (latitudes[i] > max) {
                     max = latitudes[i];
                     if (max > 0) {
-                        return max;
+                        return (int) max;
                     }   else {
                         return -1;
                     }
